+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.15.2] - 2023-08-01
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Maximal entity count in map up to 32 entities.
+
 ### Fixed
 - [@mjkuranda](https://github.com/mjkuranda): Rendering entities in `RetroRenderer`.